+++ sqry-daemon/CHANGELOG.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [9.0.11](https://github.com/verivus-oss/sqry/compare/v9.0.10...v9.0.11) - 2026-04-24
+
+### Other
+- update Cargo.lock dependencies
+## [9.0.10](https://github.com/verivus-oss/sqry/compare/v9.0.9...v9.0.10) - 2026-04-24
+
+### Other
+- update Cargo.lock dependencies
+## [9.0.9](https://github.com/verivus-oss/sqry/compare/v9.0.8...v9.0.9) - 2026-04-24
+
+### Other
+- update Cargo.lock dependencies
+## [9.0.8](https://github.com/verivus-oss/sqry/compare/v9.0.6...v9.0.8) - 2026-04-24
+
+### Fixed
+- *(daemon)* unify notify socket test guard## [9.0.7](https://github.com/verivus-oss/sqry/compare/v9.0.6...v9.0.7) - 2026-04-24
+
+### Other
+- update Cargo.lock dependencies
 ## [9.0.6](https://github.com/verivus-oss/sqry/compare/v9.0.5...v9.0.6) - 2026-04-24
 
 ### Added